+++ types/json.lua/.busted
@@ -1,0 +1,5 @@
+return {
+  _all = {
+      lpath = "../../utils/?.lua"
+  }
+}
